requirements.txt
pytest==7.3.1
selenium==4.9.0
mysql-connector-python==8.0.33
